requirements.txt
streamlit
langchain
langchain-nvidia-ai-endpoints
langchain-community
python-dotenv
faiss-cpu
pypdf